Split Flax train tests
